+++ IoVertx/pom.xml
@@ -59,7 +59,7 @@
 									implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
 									<manifestEntries>
 										<Main-Class>io.vertx.core.Starter</Main-Class>
-										<Main-Verticle>com.study.IoVertx.MyFirstVerticle</Main-Verticle>
+										<Main-Verticle>study_vertx.MyFirstVerticle</Main-Verticle>
 									</manifestEntries>
 								</transformer>
 							</transformers>
