+++ IoVertx/src/test/java/study_vertx/MyFirstVerticleTest.java
@@ -1,10 +1,7 @@
-package com.videal.ZzzVertxTest;
+package study_vertx;
 
 import io.vertx.junit5.VertxExtension;
 
-import io.vertx.junit5.VertxTestContext;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
