+++ Java8/pom.xml
@@ -27,36 +27,6 @@
 				</configuration>
 			</plugin>
 
-			<!-- a fat jar. A fat jar is a standalone executable Jar file containing 
-				all the dependencies required to run the application. This is a very convenient 
-				way to package Vert.x applications as it’s only one file. -->
-			<!-- java -jar ZzzVertxTest-1.0-SNAPSHOT-fat.jar -->
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-shade-plugin</artifactId>
-				<version>2.3</version>
-				<executions>
-					<execution>
-						<phase>install</phase>
-						<goals>
-							<goal>shade</goal>
-						</goals>
-						<configuration>
-							<transformers>
-								<transformer
-									implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-									<manifestEntries>
-										<Main-Class>io.vertx.core.Starter</Main-Class>
-										<Main-Verticle>com.videal.ZzzVertxTest.MyFirstVerticle</Main-Verticle>
-									</manifestEntries>
-								</transformer>
-							</transformers>
-							<artifactSet />
-							<outputFile>${project.build.directory}/${project.artifactId}-${project.version}-fat.jar</outputFile>
-						</configuration>
-					</execution>
-				</executions>
-			</plugin>
 		</plugins>
 	</build>
